app/requirements.txt
Flask
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-secret-manager
isodate